+++ lib/aliyun/log.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'version'
 require_relative 'log/common'
 require_relative 'log/client'
 require_relative 'log/config'